# Code to automatically generate conversion methods