Part of #6800 

There are two commits in this PR:
1. the code that I've written
2. the code that was generated - to show what code is generated.

This code is in a very early stage - the goal is to get the initial feedback, so any comments [including - this approach is bad, we should use a different one :)] are welcome.

The **initial benchmarks** show that conversion (without parsing jsons) is **~2x faster**.

cc @lavalamp @smarterclayton @fgrzadkowski @davidopp