@smarterclayton 2015-04-21
What about memory use

> On Apr 21, 2015, at 9:59 AM, Wojciech Tyczynski <notifications@github.com> wrote:
> 
> Part of #6800
> 
> There are two commits in this PR:
> 1. the code that I've written
> 2. the code that was generated (with some small manual fixes to namespaces) - to show what code is generated.
> 
> This code is in a very early stage - the goal is to get the initial feedback, so any comments [including - this approach is bad, we should use a different one :)] are welcome.
> 
> The initial benchmarks show that conversion (without parsing jsons) is ~2x faster.
> 
> cc @lavalamp @smarterclayton @fgrzadkowski @davidopp
> 
> You can view, comment on, or merge this pull request online at:
> 
>   https://github.com/GoogleCloudPlatform/kubernetes/pull/7107
> 
> Commit Summary
> 
> Generating conversion methods.
> Generated code
> File Changes
> 
> A cmd/kube-conversion/conversion.go (51)
> M pkg/api/register.go (4)
> M pkg/api/serialization_test.go (12)
> M pkg/api/v1beta3/conversion.go (431)
> A pkg/conversion/generator.go (377)
> M pkg/runtime/scheme.go (4)
> Patch Links:
> 
> https://github.com/GoogleCloudPlatform/kubernetes/pull/7107.patch
> https://github.com/GoogleCloudPlatform/kubernetes/pull/7107.diff
> —
> Reply to this email directly or view it on GitHub.
>