@wojtek-t 2015-04-21
@smarterclayton I didn't look at the memory usage yet. Do you have any specific tests on you mind? How would you like me to measure it?