@smarterclayton 2015-04-21
If you just set up a benchmark you can add --memprofile mem.out to your bench test run and then run go tool pprof ./testbinary em.out --alloc_space

----- Original Message -----
> @smarterclayton I didn't look at the memory usage yet. Do you have any
> specific tests on you mind? How would you like me to measure it?
> 
> ---
> Reply to this email directly or view it on GitHub:
> https://github.com/GoogleCloudPlatform/kubernetes/pull/7107#issuecomment-94829932