@smarterclayton 2015-04-21
I suspect we're still getting hit by the cost to generate the conversion stack and manipulate it, as well as some reflects.  Can you add a few benchmark tests to this so I can look at some profiles and we can have a standard metric?