@smarterclayton 2015-04-21
Example of benchmark: https://github.com/GoogleCloudPlatform/kubernetes/blob/98e61e252d1087648321a59f515ce0ff95f57f84/pkg/conversion/deep_copy_test.go#L113