@wojtek-t 2015-04-21
>
If you just set up a benchmark you can add --memprofile mem.out to your bench test run and then run go tool pprof ./testbinary em.out --alloc_space

Thanks - will run it tomorrow.

>
I suspect we're still getting hit by the cost to generate the conversion stack and manipulate it, as well as some reflects. 

I'm pretty sure that's true. But I think that we should start with something (the changes I've made are already giving huge improvement) and we can continue improvements after. I would like to make one step at a time.

>
Can you add a few benchmark tests to this so I can look at some profiles and we can have a standard metric?

So far I was using pkg/api/serialization_test.go. But I can add more.