@wojtek-t 2015-04-21
Sure - that's why I was so far using always pkg/api/serialization_test.go. But I agree that having more than one is good.