@smarterclayton 2015-04-21
That one will give different results every time (subtly) because, while the rand has a consistent seed, the iteration order of the runtime.Scheme map is different.  So each run is going to be generating different fuzz for different objects.

----- Original Message -----
> Sure - that's why I was so far using always pkg/api/serialization_test.go.
> But I agree that having more than one is good.
> 
> ---
> Reply to this email directly or view it on GitHub:
> https://github.com/GoogleCloudPlatform/kubernetes/pull/7107#issuecomment-94878520