@smarterclayton 2015-04-21
Yeah, good first step though.  The nice thing is you can potentially chain these and turn them into static functions with names, then have them cal each other (if necessary).

> On Apr 21, 2015, at 7:30 PM, Daniel Smith <notifications@github.com> wrote:
> 
> Hm, I was kinda thinking more along the lines of something that totally replaced the conversion functions with one big conversion function, like maybe it can take the existing conversion functions as input and glue them all together into one big function... it's interesting that this already gets a 2x speed improvement. We really want to track allocations. I think we can get this to more like 10x improvement if we cut out even more of the conversion package.
> 
> —
> Reply to this email directly or view it on GitHub.
>