@wojtek-t 2015-04-22
>
Yeah, good first step though.  The nice thing is you can potentially chain these and turn them into static functions with names, then have them cal each other (if necessary).

Yes I completely agree - this was also my plan once this is done.