@wojtek-t 2015-04-22
In #7163 I've created some dedicated benchmarks.
[result before the change vs after the change]
Pod:      384897  vs  146677
Node:    196462  vs   59733
RC:       400900  vs  133508
So the gain is more than 2.5x (on average).