@wojtek-t 2015-04-22
@smarterclayton 
I've tried to profile the memory for the benchmarks mentioned above, but either I was doing something wrong or they were completely useless [there were two boxes on the profile]. Can you please write what exactly I should do?