@smarterclayton 2015-04-22
```
GOPATH=$(pwd)/Godeps/_workspace:$GOPATH go test ./pkg/conversion -test.run=XXXXX -bench=BenchmarkDeepCopy -benchmem -memprofile mem.out -memprofilerate=1
go tool pprof --alloc_space ./conversion.test  mem.out
```

is what I use.