@smarterclayton 2015-04-22
Also, use https://github.com/golang/tools/tree/master/cmd/benchcmp to generate a diff of your benchmark results.