@wojtek-t 2015-04-23
Benchmark comparison

    benchmark                                    old ns/op     new ns/op     delta
    BenchmarkPodConversion                       388502        149268        -61.58%
    BenchmarkNodeConversion                      192973        60733         -68.53%
    BenchmarkReplicationControllerConversion     385409        135528        -64.84%

The memory allocation graphs above also show ~66% improvement.

The next step I'm going to work on is to change those conversion functions to be named and call each other without going through reflections wherever it's possible.

@smarterclayton @lavalamp 
This PR is ready for merging (although it's just the first step of improvement - more will come) [tests are passing]. Can you please merge it or do you have any additional comments now?